+++ Changelog.md
@@ -4,4 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Changed
+	
+* [DOCKER-254] Allow to mount a custom share-config-custom.xml. Location to mount: /docker-config/share-config-custom.xml.
+	
 ## [v1.0.0] Make it public